Add test categories per feature area, switch to gpt-4o-mini, refactor
