+++ tests/Batch/BatchTests.cs
@@ -14,10 +14,10 @@ namespace OpenAI.Tests.Batch;
 [TestFixture(true)]
 [TestFixture(false)]
 [Parallelizable(ParallelScope.All)]
+[Category("Batch")]
 public partial class BatchTests : SyncAsyncTestBase
 {
-    public BatchTests(bool isAsync)
-        : base(isAsync)
+    public BatchTests(bool isAsync) : base(isAsync)
     {
     }
 
@@ -55,7 +55,7 @@ public partial class BatchTests : SyncAsyncTestBase
     {
         using MemoryStream testFileStream = new();
         using StreamWriter streamWriter = new (testFileStream);
-        string input = @"{""custom_id"": ""request-1"", ""method"": ""POST"", ""url"": ""/v1/chat/completions"", ""body"": {""model"": ""gpt-3.5-turbo"", ""messages"": [{""role"": ""system"", ""content"": ""You are a helpful assistant.""}, {""role"": ""user"", ""content"": ""What is 2+2?""}]}}";
+        string input = @"{""custom_id"": ""request-1"", ""method"": ""POST"", ""url"": ""/v1/chat/completions"", ""body"": {""model"": ""gpt-4o-mini"", ""messages"": [{""role"": ""system"", ""content"": ""You are a helpful assistant.""}, {""role"": ""user"", ""content"": ""What is 2+2?""}]}}";
         streamWriter.WriteLine(input);
         streamWriter.Flush();
         testFileStream.Position = 0;
