+++ tests/Chat/ChatToolTests.cs
@@ -15,10 +15,10 @@ namespace OpenAI.Tests.Chat;
 [TestFixture(true)]
 [TestFixture(false)]
 [Parallelizable(ParallelScope.All)]
+[Category("Chat")]
 public partial class ChatToolTests : SyncAsyncTestBase
 {
-    public ChatToolTests(bool isAsync)
-        : base(isAsync)
+    public ChatToolTests(bool isAsync) : base(isAsync)
     {
     }
