+++ tests/Images/ImageGenerationTests.cs
@@ -14,15 +14,14 @@ namespace OpenAI.Tests.Images;
 [TestFixture(true)]
 [TestFixture(false)]
 [Parallelizable(ParallelScope.All)]
+[Category("Images")]
 public partial class ImageGenerationTests : SyncAsyncTestBase
 {
-    public ImageGenerationTests(bool isAsync)
-        : base(isAsync)
+    public ImageGenerationTests(bool isAsync) : base(isAsync)
     {
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task BasicGenerationWorks()
     {
         ImageClient client = GetTestClient();
@@ -40,7 +39,6 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task GenerationWithOptionsWorks()
     {
         ImageClient client = GetTestClient();
@@ -60,7 +58,6 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task GenerationWithBytesResponseWorks()
     {
         ImageClient client = GetTestClient();
@@ -82,7 +79,6 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task GenerateImageEditWorks()
     {
         ImageClient client = GetTestClient<ImageClient>(TestScenario.Images, "dall-e-2");
@@ -102,7 +98,6 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task GenerateImageEditWithMaskFileWorks()
     {
         ImageClient client = GetTestClient<ImageClient>(TestScenario.Images, "dall-e-2");
@@ -123,7 +118,6 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task GenerateImageEditWithBytesResponseWorks()
     {
         ImageClient client = GetTestClient<ImageClient>(TestScenario.Images, "dall-e-2");
@@ -146,7 +140,6 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task GenerateImageVariationWorks()
     {
         ImageClient client = GetTestClient<ImageClient>(TestScenario.Images, "dall-e-2");
@@ -164,7 +157,6 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task GenerateImageVariationWithBytesResponseWorks()
     {
         ImageClient client = GetTestClient<ImageClient>(TestScenario.Images, "dall-e-2");
@@ -186,7 +178,7 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
 
     private void ValidateGeneratedImage(Uri imageUri, string expectedSubstring)
     {
-        ChatClient chatClient = GetTestClient<ChatClient>(TestScenario.VisionChat);
+        ChatClient chatClient = GetTestClient<ChatClient>(TestScenario.Chat);
         IEnumerable<ChatMessage> messages = [
             new UserChatMessage(
                 ChatMessageContentPart.CreateTextMessageContentPart("Describe this image for me."),
@@ -200,7 +192,7 @@ public partial class ImageGenerationTests : SyncAsyncTestBase
 
     private void ValidateGeneratedImage(BinaryData imageBytes, string expectedSubstring)
     {
-        ChatClient chatClient = GetTestClient<ChatClient>(TestScenario.VisionChat);
+        ChatClient chatClient = GetTestClient<ChatClient>(TestScenario.Chat);
         IEnumerable<ChatMessage> messages = [
             new UserChatMessage(
                 ChatMessageContentPart.CreateTextMessageContentPart("Describe this image for me."),
