+++ tests/Moderations/ModerationTests.cs
@@ -9,15 +9,14 @@ namespace OpenAI.Tests.Moderations;
 [TestFixture(true)]
 [TestFixture(false)]
 [Parallelizable(ParallelScope.All)]
+[Category("Moderations")]
 public partial class ModerationTests : SyncAsyncTestBase
 {
-    public ModerationTests(bool isAsync)
-        : base(isAsync)
+    public ModerationTests(bool isAsync) : base(isAsync)
     {
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task ClassifySingleInput()
     {
         ModerationClient client = new("text-moderation-stable");
@@ -34,7 +33,6 @@ public partial class ModerationTests : SyncAsyncTestBase
     }
 
     [Test]
-    [Category("skipInCI")]
     public async Task ClassifyMultipleInputs()
     {
         ModerationClient client = new("text-moderation-stable");
